Pop-up messages added, corrected the rounding case

+++ app/src/main/java/com/example/application/models/Converter.kt
@@ -1,6 +1,8 @@
 package com.example.application.models
 
+import java.math.BigDecimal
+
 abstract class Converter {
     abstract val unitsList: List<TitleAndCode>
-    abstract fun convert(value: Double, from: String, to: String): Double
+    abstract fun convert(value: BigDecimal, from: String, to: String): BigDecimal
 }
